feat(boss/v2): Add interval and status on RegisterTask
This allows us to populate those fields on the BOSS server right when
creating a new task.

Regarding the implementation on the BOSS CLI, we can probably set the
interval to 0 by default until we know of a game that uses it. The
`status` can either be `open` or `close`, we can probably make `open`
the default here.